+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/review/application/usecase/ReviewLikeUseCase.java
@@ -1,0 +1,5 @@
+package com.romanticpipe.reviewcanvas.domain.review.application.usecase;
+
+public interface ReviewLikeUseCase {
+	int getReviewLikeCount(Long reviewId);
+}
