+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/review/presentation/v1/ReviewApi.java
@@ -2,7 +2,6 @@ package com.romanticpipe.reviewcanvas.domain.review.presentation.v1;
 
 import java.util.EnumSet;
 import java.util.List;
-import java.util.Map;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -220,14 +219,4 @@ interface ReviewApi {
 		@RequestPart(required = false) List<MultipartFile> reviewImages
 	);
 
-	@Operation(summary = "리뷰 좋아요 개수 조회 API", description = "특정 리뷰의 좋아요 개수를 조회한다.")
-	@ApiResponses(value = {
-		@ApiResponse(
-			responseCode = "200",
-			description = "성공적으로 리뷰 좋아요 개수 조회가 완료되었습니다.")
-	})
-	@GetMapping("/reviews/{reviewId}/like/count")
-	ResponseEntity<SuccessResponse<Map<String, Integer>>> getReviewLikeCount(
-		@PathVariable("reviewId") Long reviewId
-	);
 }
