+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/review/presentation/v1/ReviewController.java
@@ -3,7 +3,6 @@ package com.romanticpipe.reviewcanvas.domain.review.presentation.v1;
 import java.time.LocalDateTime;
 import java.util.EnumSet;
 import java.util.List;
-import java.util.Map;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -202,13 +201,4 @@ class ReviewController implements ReviewApi {
 		return SuccessResponse.ofNoData().asHttp(HttpStatus.OK);
 	}
 
-	@GetMapping("/reviews/{reviewId}/like/count")
-	public ResponseEntity<SuccessResponse<Map<String, Integer>>> getReviewLikeCount(
-		@PathVariable("reviewId") Long reviewId
-	) {
-		int likeCount = reviewUseCase.getReviewLikeCount(reviewId);
-		return SuccessResponse.of(
-			Map.of("count", likeCount)
-		).asHttp(HttpStatus.OK);
-	}
 }
