+++ domain-module/review/src/main/java/com/romanticpipe/reviewcanvas/service/ReviewLikeService.java
@@ -1,0 +1,18 @@
+package com.romanticpipe.reviewcanvas.service;
+
+import org.springframework.stereotype.Service;
+
+import com.romanticpipe.reviewcanvas.repository.ReviewLikeRepository;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class ReviewLikeService {
+
+	private final ReviewLikeRepository reviewLikeRepository;
+
+	public int getReviewLikeCount(Long reviewId) {
+		return 0;
+	}
+}
