+++ domain-module/review/src/main/java/com/romanticpipe/reviewcanvas/service/ReviewService.java
@@ -81,5 +81,4 @@ public class ReviewService {
 				replyFilters)
 		);
 	}
-
 }
